+++ cmd/list_cmd.go
@@ -23,7 +23,6 @@ import (
 	"reflect"
 	"sort"
 
-	log "github.com/sirupsen/logrus"
 	"github.com/synfinatic/aws-sso-cli/utils"
 	"github.com/synfinatic/gotable"
 )